+++ aspnetcore/tutorials/min-web-api/samples/7.x/todo/Todo.cs
@@ -1,0 +1,6 @@
+class Todo
+{
+    public int Id { get; set; }
+    public string? Name { get; set; }
+    public bool IsComplete { get; set; }
+}
