+++ aspnetcore/tutorials/min-web-api/samples/7.x/todoDTO/Program.cs
@@ -1,5 +1,3 @@
-#define FIRST // FIRST FINAL
-#if FIRST
 // <snippet_all>
 using Microsoft.EntityFrameworkCore;
 
@@ -7,8 +5,6 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddDbContext<TodoDb>(opt => opt.UseInMemoryDatabase("TodoList"));
 builder.Services.AddDatabaseDeveloperPageExceptionFilter();
 var app = builder.Build();
-
-app.MapGet("/", () => "Hello World!");
 
 // <snippet_group>
 RouteGroupBuilder group = app.MapGroup("/todoitems");
@@ -78,36 +74,4 @@ static async Task<IResult> DeleteTodo(int id, TodoDb db)
     return TypedResults.NotFound();
 }
 // <snippet_handlers>
-
-//<snippet_secret>
-public class Todo
-{
-    public int Id { get; set; }
-    public string? Name { get; set; }
-    public bool IsComplete { get; set; }
-    public string? Secret { get; set; }
-}
-//</snippet_secret>
-
-// <snippet_DTO>
-public class TodoItemDTO
-{
-    public int Id { get; set; }
-    public string? Name { get; set; }
-    public bool IsComplete { get; set; }
-
-    public TodoItemDTO() { }
-    public TodoItemDTO(Todo todoItem) =>
-    (Id, Name, IsComplete) = (todoItem.Id, todoItem.Name, todoItem.IsComplete);
-}
-// </snippet_DTO>
-
-class TodoDb : DbContext
-{
-    public TodoDb(DbContextOptions<TodoDb> options)
-        : base(options) { }
-
-    public DbSet<Todo> Todos => Set<Todo>();
-}
-#endif
 // </snippet_all>
