+++ aspnetcore/tutorials/min-web-api/samples/7.x/todoDTO/TodoItemDTO.cs
@@ -1,0 +1,10 @@
+public class TodoItemDTO
+{
+    public int Id { get; set; }
+    public string? Name { get; set; }
+    public bool IsComplete { get; set; }
+
+    public TodoItemDTO() { }
+    public TodoItemDTO(Todo todoItem) =>
+    (Id, Name, IsComplete) = (todoItem.Id, todoItem.Name, todoItem.IsComplete);
+}
