+++ aspnetcore/tutorials/min-web-api/samples/7.x/todoTypedResults/TodoDb.cs
@@ -1,0 +1,9 @@
+using Microsoft.EntityFrameworkCore;
+
+class TodoDb : DbContext
+{
+    public TodoDb(DbContextOptions<TodoDb> options)
+        : base(options) { }
+
+    public DbSet<Todo> Todos => Set<Todo>();
+}
